20932 bump dina base 0.42
bump version and adjust CFRs
