+++ pom.xml
@@ -37,7 +37,7 @@
 		<asciidoctor-maven-plugin.version>2.1.0</asciidoctor-maven-plugin.version>
 		<asciidoctorj.diagram.version>2.0.2</asciidoctorj.diagram.version>
 		<hibernate-types-52.version>2.9.13</hibernate-types-52.version>
-		<dina-base-api.version>0.41</dina-base-api.version>
+		<dina-base-api.version>0.42</dina-base-api.version>
 
 		<!-- Override to avoid NoClassDefFoundError: org/apache/maven/doxia/siterenderer/DocumentContent -->
 		<maven-site-plugin.version>3.7.1</maven-site-plugin.version>
